split up into quick and index, make sort requireable
