+++ lib/keys/index.js
@@ -12,6 +12,7 @@ exports.properties = {
   keyType: true,
   order: true, // make this a key
   sort (val) {
+    console.log('lets sort!')
     // and do shit!
     this.sort = val
   }
@@ -49,7 +50,6 @@ exports.define = {
         addFilters(this, key)
       }
     }
-    // reverse qs
     if ('sort' in this) {
       const sort = this.sort
       if (sort) {
@@ -71,6 +71,7 @@ exports.define = {
           for (let i = 0; i < len; i++) {
             if (keys[i] === key) {
               keys.splice(i, 1)
+              if (keys._) { keys._.splice(i, 1) }
               break
             }
           }
@@ -93,6 +94,7 @@ exports.define = {
         for (let i = 0; i < len; i++) {
           if (keys[i] === key) {
             keys.splice(i, 1)
+            if (keys._) { keys._.splice(i, 1) }
             break
           }
         }
