+++ lib/keys/sort/index.js
@@ -1,8 +1,13 @@
 'use strict'
+const quickSort = require('./quick')
 
-global.cnt = 0
+module.exports = exports = function (target, keys, left, right) {
+  if (!right) { right = keys.length }
+  if (!left) { left = 0 }
+  quickSort(target, keys, left, right)
+}
+
 exports.add = function (parent, target, arr, field) {
-  // combine creation for extra speed
   const len = arr.length
   var order = field in target && target[field]
   if (order) {
@@ -17,6 +22,7 @@ exports.add = function (parent, target, arr, field) {
   if (len > 1) {
     const index = len - 1
     // update needs to find it self as well
+    // try to reuse a bit in update
     for (let i = len - 2; i > -1; i--) {
       if (order > orders[i]) {
         break
@@ -39,45 +45,6 @@ exports.add = function (parent, target, arr, field) {
   }
 }
 
-exports.remove = function (index, arr) {
-  if (arr._) {
-    arr._.splice(index, 1)
-  }
-}
-
 exports.update = function update (parent, target, arr) {
   console.log('update!')
-}
-
-// quick is only for total sort (when you add field ordered)
-// lets renome to "sort" // true is only thing that is supported later add method
-exports.quick = function quick (target, arr, left, right) {
-  var pivot, index
-  if (left < right) {
-    pivot = right
-    index = check(target, arr, pivot, left, right)
-    quick(target, arr, left, index - 1)
-    quick(target, arr, index + 1, right)
-  }
-  return arr
-}
-
-function check (target, arr, pivot, left, right) {
-  const pivotValue = target[arr[pivot]].order
-  var index = left
-  for (let i = left; i < right; i++) {
-    // dont do this fill in ._ array -- if its there fill it as well
-    if (target[arr[i]].order < pivotValue) {
-      swap(arr, i, index)
-      index++
-    }
-  }
-  swap(arr, right, index)
-  return index
-}
-
-function swap (arr, i, j) {
-  const temp = arr[i]
-  arr[i] = arr[j]
-  arr[j] = temp
 }
